Updated properties and mpi_interfaces
