+++ include/mesh/mpi_interfaces/mpi_interfaces.hpp
@@ -1,0 +1,24 @@
+#ifndef _MPI_INTERFACES_HPP
+#define _MPI_INTERFACES_HPP
+
+#include "kokkos_abstractions.h"
+#include "specfem_mpi.h"
+
+namespace specfem {
+namespace mesh {
+namespace interfaces {
+
+struct interface {
+  // Utilities use to compute MPI buffers
+  int ninterfaces, max_interface_size;
+  specfem::kokkos::HostView1d<int> my_neighbors, my_nelmnts_neighbors;
+  specfem::kokkos::HostView3d<int> my_interfaces;
+  interface(){};
+  interface(const int ninterfaces, const int max_interface_size);
+  interface(std::ifstream &stream, const specfem::MPI::MPI *mpi);
+};
+} // namespace interfaces
+} // namespace mesh
+} // namespace specfem
+
+#endif
